+++ src/app/(auth)/login/social/kakao/page.tsx
@@ -1,36 +1,16 @@
 'use client';
 import LoadingSpinner from '@/src/components/primitives/LoadingSpinner';
 import { KAKAO_REDIRECT_URI_LOGIN } from '@/src/constants/social';
-import {
-  KakaoLoginRequestBody,
-  kakaoLoginUser,
-} from '@/src/services/pages/login/api';
-import { queries } from '@/src/services/primitives/queries';
-import { setTokenAction } from '@/src/services/primitives/tokenAction';
-import { TokenUserResponseType } from '@/src/types/userType';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import useKakaoLoginUser from '@/src/hooks/auth/useKakaoLoginUser';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect, useRef } from 'react';
 
 export default function KakaoLoginPage() {
-  const queryClient = useQueryClient();
   const router = useRouter();
   const flagRef = useRef(false);
   const searchParams = useSearchParams();
   const code = searchParams.get('code');
-  const kakaoLoginMutation = useMutation({
-    mutationFn: async (data: KakaoLoginRequestBody) =>
-      await kakaoLoginUser(data),
-    onSuccess: async (data: TokenUserResponseType) => {
-      queryClient.setQueryData(queries.user(), data.user); // 리액트 쿼리 데이터 캐싱
-      await setTokenAction(data); // 토큰 쿠키 설정
-      router.replace('/');
-    },
-    onError: () => {
-      alert('로그인에 실패하였습니다. 다시 시도 해주세요.');
-      router.replace('/login');
-    },
-  });
+  const kakaoLoginMutation = useKakaoLoginUser();
 
   useEffect(() => {
     if (!code) {
